+++ system_client/src/model/location.hpp
@@ -7,8 +7,8 @@ class Location
 {
 	private:
 		uint32_t id;
-		uint32_t x;
-		uint32_t y;
+		double x;
+		double y;
 		double a;
 		std::string description; 
 		bool isDepot;
@@ -16,15 +16,15 @@ class Location
 	public:
 		Location(uint32_t id = 0, uint32_t x = 0, uint32_t y = 0, std::string description = "Empty", bool isDepot = false);
 		void setId(uint32_t id);
-		void setX(uint32_t x);
-		void setY(uint32_t y);
+		void setX(double x);
+		void setY(double y);
 		void setA(double a);
 		void setDescription(std::string description);
 		void setIsDepot(bool isDepot);
 
 		uint32_t getId();
-		uint32_t getX();
-		uint32_t getY();
+		double getX();
+		double getY();
 		double getA();
 		std::string getDescription();
 		bool getIsDepot();
